fix bookmark tree
git-svn-id: http://svn.ilias.de/svn/ilias/trunk@24589 21b2c9ec-7c21-0410-8b45-9bfb8ed2bfc5

+++ Services/PersonalDesktop/classes/class.ilBookmark.php
@@ -93,10 +93,13 @@ class ilBookmark
 	function delete()
 	{
 		global $ilDB;
-		
-		$q = "DELETE FROM bookmark_data WHERE obj_id = ".
-			$ilDB->quote($this->getId(), "integer");
-		$ilDB->manipulate($q);
+
+		if ($this->getId() != 1)
+		{
+			$q = "DELETE FROM bookmark_data WHERE obj_id = ".
+				$ilDB->quote($this->getId(), "integer");
+			$ilDB->manipulate($q);
+		}
 	}
